Wait to Start, Cleanup
I made the game wait to start until the user presses Enter. I also
edited the movement limit slightly, so the player can get a little
closer to the edge of the screen. It needs to be edited more, but
cropping empty space off of the images might help, too. Also, I removed
the license headers from the source files, since we're not using them.
Finally, we ended up with some file duplication in different
directories; I deleted the old ones, to keep track of which files are
the latest ones.